\docType{import}
\name{reexports}
\alias{reexports}
\alias{reexported_fn}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{otherpkg}{\code{\link[otherpkg]{reexported_fn}}}
}}
